+++ services/summarizer/TOC_generator.py
@@ -18,8 +18,8 @@ from dataclasses import dataclass
 from pathlib import Path
 
 # Import enhanced components
-from services.token_manager import TokenManager, create_token_manager
-from services.textrank_summarizer import HybridSummarizerStrategy, create_hybrid_summarizer
+from services.summarizer.token_manager import TokenManager, create_token_manager
+from services.summarizer.textrank_summarizer import HybridSummarizerStrategy, create_hybrid_summarizer
 
 logger = logging.getLogger(__name__)
